+++ R/promise.R
@@ -1,3 +1,0 @@
-promise_was_forced <- function(name, env = cartographer_global) {
-  .Call(C_promise_was_forced, name, env)
-}