+++ src/cartographer.h
@@ -1,3 +1,0 @@
-#include <Rinternals.h>
-
-SEXP promise_was_forced(SEXP name, SEXP env);